github: Implement workflows and README.
This commit will always be one ahead of pkgsrc trunk.  It imports the
necessary workflows into the ~/.github directory, and updates the README
with instructions on how to use this repository.

To keep things simple and tidy, any changes to these scripts will be
squashed and rebased, so this commit will be the only one to include all
of the changes.